Populate serviceAccount based upon annotation
The com.openfaas.serviceaccount annotation is used from the
function spec to assign a serviceAccount to the podSpec.

Tested on Kind with Kubernetes 1.13. First by setting the value
as build-robot (as per README.md) and seeing it applied on the
podSpec, then by removing the annotation and seeing it go back
to "default".

Signed-off-by: Alex Ellis <alexellis2@gmail.com>