Add logs feature and namespace dummy endpoint
Adds logs feature from faas-netes so that users can stream logs
from containers. Closes: #99

Partial fix for: #92 which adds a dummy /system/namespaces
endpoint which returns the namespace set at configuration time.
Further work will be required to support CRD reconcilation into
multiple namespaces.

Tested with Kubernetes v1.14.1 on a single-node cluster using
image: openfaas/openfaas-operator:0.9.10-rc1

The RBAC roles also had to be updated in the helm chart in
faas-netes to allow for the logs to be viewed and for pods to be
listed.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>